+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v3.0.0
+- added basic accounts support
+- `notifConnector` method now always use smart-notifier/kraken format
+- upgraded `postConnector` method to send connector data (this affects also triggering a batch extract)
+
 # v2.1.2
 - added `stubUsersSegments` and `stubAccountsSegments` for separate segments definition
 - marked `stubSegments` as deprecated